refactor: hooks 폴더내 notice 및 timer 서브 폴더 생성
