+++ src/hooks/study-detail/timers/useTimers.ts
@@ -1,7 +1,7 @@
 import useGetTimers from './useGetTimers';
 import { useState, useEffect } from 'react';
-import { TimerType } from '../../types/interface';
-import studyMainApi from '../../api/studyMainApi';
+import { TimerType } from '../../../types/interface';
+import studyMainApi from '../../../api/studyMainApi';
 import * as Sentry from '@sentry/react';
 
 export default function useTimers(studyId: number) {
@@ -33,10 +33,10 @@ export default function useTimers(studyId: number) {
           prev.map((clientTimer) => {
             const serverTimer = data.find((s: TimerType) => s.userId === clientTimer.userId);
 
-            if (!serverTimer) return clientTimer;
-
-            if (Math.abs(serverTimer.timerTime - clientTimer.timerTime) > 1) {
-              return { ...clientTimer, timerTime: serverTimer.timerTime };
+            if (serverTimer) {
+              if (Math.abs(serverTimer.timerTime - clientTimer.timerTime) > 1) {
+                return { ...clientTimer, timerTime: serverTimer.timerTime };
+              }
             }
 
             return clientTimer;
